refactor profile card component
