+++ web/src/pages/people.js
@@ -3,7 +3,7 @@ import { jsx } from "theme-ui";
 import { graphql } from "gatsby";
 import Container from "../components/core/container";
 import GraphQLErrorList from "../components/core/graphql-error-list";
-import PeopleList from "../components/people-list";
+import PeopleList from "../components/people-layouts/people-list";
 import SEO from "../components/core/seo";
 import Layout from "../containers/layout";
 import { mapEdgesToNodes, filterOutDocsWithoutSlugs } from "../lib/helpers";
@@ -14,6 +14,7 @@ export const query = graphql`
       filter: { 
         position: { title: { eq: "Contributor" } }
       }
+      sort: { fields: [name], order: ASC }
     )
     {
       edges {
